# Optimise the single-clause chsend and chrecv