I've significantly sped up the `chsend` and `chrecv` implementations. I believe this fits in the philosophy of avoiding `choose` , for reference: [Select Statement Considered Harmful](http://250bpm.com/blog:72).

- `hquery` gets called thrice per clause in `choose`, I've implemented last-call caching in `hquery`.
- `chsend` and `chrecv` only need to call `hquery` once
- The initialisation of the clause structure fills `reserved` with zeroes causing an expensive `rep stos` on x86 architectures. Only filling in the necessary elements makes this much faster.
- `chsend` and `chrecv` can avoid setting/checking `errno` until there really is an error; setting `errno` is typically expensive.

On a i7 4770k (GCC 4.9.4), previously:
```bash
$ ./perf/chan 10
done 10M roundtrips in 1.666000 seconds
duration of passing a single message: 83 ns
message passes per second: 12.048192M
```

Afterwards:
```bash
$ ./perf/chan 10
done 10M roundtrips in 0.815000 seconds
duration of passing a single message: 40 ns
message passes per second: 25.000000M
```

Statically compiled with heavier optimisations in one line:
```C
$ gcc -march=native -O3 -flto -fvisibility=hidden *.c perf/chan.c
$ ./a.out 10
done 10M roundtrips in 0.538000 seconds
duration of passing a single message: 26 ns
message passes per second: 38.461536M
```