@sustrik 2016-11-18
Btw, here's a patch to simplify chsend and chrecv: https://github.com/sustrik/libdill/commit/ad7eece0dde4dc1f3ab9a6cb79d6454e595fab70

I hope it doesn't break your careful optimization. Can you give it a look?