@raedwulf 2016-11-18
Seems good, the logic looks sound - and is a couple of nanoseconds faster on all my systems (probably due to the `dill_assert` being dropped).